+++ RELEASE_NOTES.md
@@ -23,3 +23,6 @@
 - Multiscreen levels
 - New level: The Deep
 - New level: Abyss Hunter
+- Gamma control
+- Change logo
+- Removed all AI art